+++ src/DateRange.ts
@@ -8,25 +8,36 @@ export class EndDateMustBeOnOrAfterStartDateError extends Error {
   }
 }
 
+type DateRangeParams = {
+  start?: LocalDate;
+  end?: LocalDate;
+};
+
 export class DateRange {
-  constructor(readonly start: LocalDate, readonly end?: LocalDate) {
-    if (end?.isBefore(start)) {
+  readonly start?: LocalDate;
+
+  readonly end?: LocalDate;
+
+  constructor(params: DateRangeParams) {
+    if (endDateBeforeStartDate(params)) {
       throw new EndDateMustBeOnOrAfterStartDateError();
     }
+    this.start = params?.start;
+    this.end = params?.end;
   }
 
   /**
    * Returns `undefined` if date range args are invalid (that is, when `end` is before `start`).
    */
-  static from(start: LocalDate, end?: LocalDate): DateRange | undefined {
-    if (end?.isBefore(start)) {
+  static from(params: DateRangeParams): DateRange | undefined {
+    if (endDateBeforeStartDate(params)) {
       return undefined;
     }
-    return new DateRange(start, end);
+    return new DateRange(params);
   }
 
   static compare(a: DateRange, b: DateRange): number {
-    if (a.start.value === b.start.value) {
+    if (a.start?.value === b.start?.value) {
       if (a.end && b.end) {
         return LocalDate.compare(a.end, b.end);
       }
@@ -39,11 +50,19 @@ export class DateRange {
       return 0;
     }
 
+    if (!a.start) {
+      return -1;
+    }
+
+    if (!b.start) {
+      return 1;
+    }
+
     return LocalDate.compare(a.start, b.start);
   }
 
   getCurrentness(asOf: LocalDate = LocalDate.today()): DateRangeCurrentness {
-    if (this.start.isAfter(asOf)) {
+    if (this.start?.isAfter(asOf)) {
       return "future";
     }
 
@@ -67,12 +86,27 @@ export class DateRange {
   }
 
   format({ dateFormat }: { dateFormat?: LocalDateFormat } = {}): string {
-    if (this.end) {
+    if (this.start && this.end) {
       return `${this.start.format(dateFormat)} – ${this.end.format(
         dateFormat
       )}`;
     }
 
-    return `From ${this.start.format(dateFormat)}`;
+    if (this.start) {
+      return `From ${this.start.format(dateFormat)}`;
+    }
+
+    if (this.end) {
+      return `Until ${this.end.format(dateFormat)}`;
+    }
+
+    return `Forever`;
   }
+}
+
+function endDateBeforeStartDate(params: DateRangeParams): boolean {
+  if (!params.start || !params.end) {
+    return false;
+  }
+  return params.end.isBefore(params.start);
 }