+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### ⚠ BREAKING CHANGES
+
+* **DateRange:** `DateRange can now be constructed with no start date, and is now instantiated with an object {start?: LocalDate, end?: LocalDate}`
+
 ### [2.12.1](https://github.com/ailohq/date/compare/v2.12.0...v2.12.1) (2021-08-15)
 
 ## [2.12.0](https://github.com/ailohq/date/compare/v2.11.0...v2.12.0) (2021-08-13)